Headband Prototype that includes use of Pictures
This script will place an image on a user forehead and follow the user's movement. The image used for this script is not uploaded so please change the image path in the code to match an image that you will be using. Images will be uploaded when we organize the repository this week.